backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23
python-socketio>=5.10
pydantic>=2.3,<3.0
starlette>=0.37

